Change definition of reset_time to floating point
Truncating reset_time (or start of next window) by
casting to an int was incorrectly providing the wrong
reset_time as there was precision loss in the conversion.

+++ limits/aio/storage/memcached.py
@@ -126,14 +126,14 @@ class MemcachedStorage(Storage):
 
         return amount
 
-    async def get_expiry(self, key: str) -> int:
+    async def get_expiry(self, key: str) -> float:
         """
         :param key: the key to get the expiry for
         """
         storage = await self.get_storage()
         item = await storage.get(f"{key}/expires".encode())
 
-        return int(item and float(item.value) or time.time())
+        return item and float(item.value) or time.time()
 
     async def check(self) -> bool:
         """
